Move HelloWorld extension to subpackage
