+++ build.gradle
@@ -27,8 +27,8 @@ apply plugin: 'application'
 apply from: 'gradle/formatting.gradle'
 apply plugin: 'maven-publish'
 
-mainClassName = 'org.opensearch.sdk.sample.HelloWorldExtension'
-
+mainClassName = 'org.opensearch.sdk.ExtensionsRunner'
+  
 group 'org.opensearch.sdk'
 version '1.0.0-SNAPSHOT'
 
@@ -98,12 +98,16 @@ task javadocStrict(type: Javadoc) {
   classpath = sourceSets.main.runtimeClasspath
   options.addStringOption('Xdoclint:all', '-quiet')
   options.memberLevel = JavadocMemberLevel.PRIVATE
-  
-  // the netty4 package will eventually be published to mavenCentral
-  // See https://github.com/opensearch-project/OpenSearch/issues/3118
-  exclude 'org/opensearch/sdk/netty4'
 }
 check.dependsOn javadocStrict
+
+// this task runs the helloworld sample extension
+task helloWorld(type: JavaExec) {
+  group = 'Execution'
+  description = 'Run HelloWorld Extension.'
+  mainClass = 'org.opensearch.sdk.sample.helloworld.HelloWorldExtension'
+  classpath = sourceSets.main.runtimeClasspath
+}
 
 test {
     useJUnitPlatform()
