+++ src/test/java/org/opensearch/sdk/sample/helloworld/rest/TestRestHelloAction.java
@@ -5,7 +5,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.sdk.sample.rest;
+package org.opensearch.sdk.sample.helloworld.rest;
 
 import java.nio.charset.StandardCharsets;
 import java.util.List;
